+++ src/app/(dashboard)/items/page.tsx
@@ -11,7 +11,7 @@ import StatusTag, { getStockStatus } from "@/components/features/dashboard/atoms
 import { PencilSquareIcon, TrashIcon, MagnifyingGlassIcon } from "@heroicons/react/24/solid";
 import DashboardContentLayout from "@/components/features/dashboard/templates/DashboardContentLayout";
 import { ColumnDef } from "@tanstack/react-table";
-import { Item, useProducts } from "@/hook/dashboard/useProduct";
+import { Item, useProducts } from "@/hook/dashboard/items/useProduct";
 import { BaseTable } from "@/components/features/dashboard/atoms/BaseTable";
 import Pagination from "@/components/atoms/Pagination";
